1.fix query group by userid 2.add feature about adding user into group

+++ src/main/java/com/tencent/wxcloudrun/dao/GroupMapper.java
@@ -11,7 +11,7 @@ public interface GroupMapper {
 
     int insert(Group record);
 
-    Long insertSelective(Group record);
+    int insertSelective(Group record);
 
     Group selectByPrimaryKey(Long id);
